requirements.txt
streamlit
pdfplumber
matplotlib
plotly
pillow